requirements.txt
absl-py==0.9.0
alembic==1.4.2
APScheduler==3.6.3
astor==0.8.1
bleach==1.5.0
captcha==0.3
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
croniter==0.3.31
cycler==0.10.0
easydict==1.8
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Migrate==2.5.3
Flask-RQ2==18.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gast==0.3.3
grpcio==1.27.2
html5lib==0.9999999
idna==2.7
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==2.1.0
numpy==1.16.2
olefile==0.46
Pillow==7.0.0
protobuf==3.11.3
psutil==5.7.0
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
redis==3.4.1
requests==2.19.1
rq==1.3.0
rq-scheduler==0.9.1
six==1.14.0
SQLAlchemy==1.3.15
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
tzlocal==2.0.0
urllib3==1.23
Werkzeug==1.0.0
WTForms==2.2.1
